requirements.txt
feedparser
prefect==2.20.18
psycopg2-binary
python-dotenv
openai